builds again

+++ sim2/tank.h
@@ -14,4 +14,4 @@ double distanceto(tank v, double x, double y);
 void driveto(tank v, double x, double y);
 void b_driveto(tank *v, double x, double y, double r);
 void drivetogoal(tank *v);
-void rotate(tank *v, int dh);
+void rotate(tank *v, double dh);
